Добавил механизм доступа пользователя только к своим карточкам, проверку принадлежности редактируемой/удаляемой карточки активному пользователю, контроль и изменение счетчика карточек пользователя в том числе ограничение по количеству карточек

+++ app/Http/Controllers/AjaxController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request; // для работы с Request
 use App\Card; // Использовать пространство имен модели Book
+use Illuminate\Support\Facades\Auth;
+use phpDocumentor\Reflection\Types\False_;
 
 class AjaxController extends Controller
 {
@@ -22,11 +24,15 @@ class AjaxController extends Controller
      * 
      * @return void;
      * 
-     * Printss array of cards - JSON.
+     * Prints array of cards - JSON.
      */
     public function getCards()
     {
 
+        // Получаем ID аутентифицированного пользователя
+        $userId = Auth::id();
+
+        // Получаем карточки, принадлежащие пользователю
         $cards = Card::select(
             'id',
             'original',
@@ -35,7 +41,7 @@ class AjaxController extends Controller
             'translationComment',
             'stack'
         )
-            ->where('login', 'chistowick')
+            ->where('login', $userId)
             ->get();
 
         // Converts the array to JSON and returns it to js-client
@@ -52,18 +58,37 @@ class AjaxController extends Controller
      */
     public function addCard(Request $request)
     {
+        $level = Auth::user()->level;
+        $max_level = Auth::user()->max_level;
 
-        $newCard = Card::create(
-            [
-                'original'              => $request->original,
-                'originalComment'       => $request->originalComment,
-                'translation'           => $request->translation,
-                'translationComment'    => $request->translationComment,
-                'stack'                 => $request->stack,
-            ]
-        );
+        // Убеждаемся, что максимум карточек для пользователя не достигнут
+        if ($level < $max_level) {
 
-        echo json_encode($newCard->id);
+            // Получаем ID аутентифицированного пользователя
+            $userId = Auth::id();
+
+            $newCard = Card::create(
+                [
+                    'original'              => $request->original,
+                    'originalComment'       => $request->originalComment,
+                    'translation'           => $request->translation,
+                    'translationComment'    => $request->translationComment,
+                    'stack'                 => $request->stack,
+                    'login'                 => $userId,
+                ]
+            );
+
+            // Отправляем скрипту JS id созданной карточки для создания карточки на месте
+            echo json_encode($newCard->id);
+
+            // Добавляем "1" к счетчику заполнения базы
+            Auth::user()->level = ($level + 1);
+            Auth::user()->save();
+        } else {
+
+            // Сообщаем клиенту, что для него достигнуто максимальное количество карточек 
+            echo json_encode(['overload']);
+        }
 
         return;
     }
@@ -77,25 +102,31 @@ class AjaxController extends Controller
      **/
     public function editCard(Request $request)
     {
+        // Получаем ID аутентифицированного пользователя
+        $userId = Auth::id();
+
+        // Извлекаем ID редактируемой карточки, из клиентского запроса
         $editedCardId = $request->id;
 
-        // TODO Нужно добавить проверку логина из сессии, чтобы пользователь мог редактировать только свои карточки
+        //Получаем модель карточки, которую хотим отредактировать
+        $editedCard = Card::find($editedCardId);
 
-        $editedCard = Card::find($editedCardId); //Получаем модель карточки, которую хотим отредактировать
+        // Если ID аутентифицированного пользователя совпадает с login владельца карточки, позволяем редактирование
+        if ($editedCard->login == $userId) {
 
-        // Изменяем карточку
-        $editedCard->original           = $request->original;
-        $editedCard->originalComment    = $request->originalComment;
-        $editedCard->translation        = $request->translation;
-        $editedCard->translationComment = $request->translationComment;
-        $editedCard->stack              = $request->stack;
+            // Изменяем карточку
+            $editedCard->original           = $request->original;
+            $editedCard->originalComment    = $request->originalComment;
+            $editedCard->translation        = $request->translation;
+            $editedCard->translationComment = $request->translationComment;
+            $editedCard->stack              = $request->stack;
 
-        // Сохраняем изменения в базу и возвращаем статус операции клиенту JS
-        // TODO Проверить тип данных, в котором TRUE/FALSE придет в JS
-        if ($editedCard->save()) {
-            echo json_encode([TRUE]);
-        } else {
-            echo json_encode([FALSE]);
+            // Сохраняем изменения в базу и возвращаем статус операции клиенту JS
+            if ($editedCard->save()) {
+                echo json_encode([TRUE]);
+            } else {
+                echo json_encode([FALSE]);
+            }
         }
 
         return;
@@ -110,15 +141,35 @@ class AjaxController extends Controller
      **/
     public function deleteCard(Request $request)
     {
-        $deletedCardId = $request->id;
+        $level = Auth::user()->level;
 
-        // TODO Нужно добавить проверку логина из сессии, чтобы пользователь мог редактировать только свои карточки
+        // Убеждаемся, что у пользователя есть карточки
+        if ($level > 0) {
 
-        // Удаляем модель по id, переданному от JS клиента и возвращаем статус операции
-        if (Card::destroy($deletedCardId)) {
-            echo json_encode([TRUE]);
-        } else {
-            echo json_encode([FALSE]);
+            // Получаем ID аутентифицированного пользователя
+            $userId = Auth::id();
+
+            // Извлекаем ID удаляемой карточки, из клиентского запроса
+            $deletedCardId = $request->id;
+
+            //Получаем модель карточки, которую хотим удалить
+            $deletedCard = Card::find($deletedCardId);
+
+            // Если ID аутентифицированного пользователя совпадает с login владельца карточки, позволяем удаление
+            if ($deletedCard->login == $userId) {
+
+                // Удаляем модель по id, переданному от JS клиента и возвращаем статус операции
+                // TODO Проверить тип данных, в котором TRUE/FALSE придет в JS
+                if (Card::destroy($deletedCardId)) {
+                    echo json_encode([TRUE]);
+                } else {
+                    echo json_encode([FALSE]);
+                }
+            }
+
+            // Уменьшаем счетчик карточек пользователя на "1"
+            Auth::user()->level = ($level - 1);
+            Auth::user()->save();
         }
 
         return;
@@ -133,21 +184,28 @@ class AjaxController extends Controller
      **/
     public function moveCard(Request $request)
     {
+        // Получаем ID аутентифицированного пользователя
+        $userId = Auth::id();
+
+        // Извлекаем ID перемещаемой карточки, из клиентского запроса
         $movedCardId = $request->id;
 
-        // TODO Нужно добавить проверку логина из сессии, чтобы пользователь мог редактировать только свои карточки
+        // Получаем модель карточки, которую хотим отредактировать
+        $movedCard = Card::find($movedCardId);
 
-        $movedCard = Card::find($movedCardId); // Получаем модель карточки, которую хотим отредактировать
+        // Если ID аутентифицированного пользователя совпадает с login владельца карточки, позволяем перемещение
+        if ($movedCard->login == $userId) {
 
-        // Изменяем стек карточки
-        $movedCard->stack = $request->stack;
+            // Изменяем стек карточки
+            $movedCard->stack = $request->stack;
 
-        // Сохраняем изменения в базу и возвращаем статус операции клиенту JS
-        // TODO Проверить тип данных, в котором TRUE/FALSE придет в JS
-        if ($movedCard->save()) {
-            echo json_encode([TRUE]);
-        } else {
-            echo json_encode([FALSE]);
+            // Сохраняем изменения в базу и возвращаем статус операции клиенту JS
+            // TODO Проверить тип данных, в котором TRUE/FALSE придет в JS
+            if ($movedCard->save()) {
+                echo json_encode([TRUE]);
+            } else {
+                echo json_encode([FALSE]);
+            }
         }
 
         return;
